Solve part 2 of day 5
This implementation is pretty slow, but still fast enough not to require any
optimization with the given input. A quick optimization would be to also
calculate the list length as part of `reduce` and then compare by length,
instead of checking list equality to determine if the result has changed.