Actually solve part 2 of day 19
The program was too slow to run, but eventually solved it by writing
`sum(divisors 10551376)` into wolfram alpha. By monitoring various operations
it was doing, it became apparent, that it essentially slowly increments values
in the 2nd register and adds the value to the first register, if it's a divisor
of the constant 3rd register `10551376`.